Make user unable to make account with existing username

+++ js/signup.js
@@ -9,21 +9,34 @@ signupBtn.addEventListener('click', (event) => {
   if (usernameInput.value && passwordInput.value) {
     let users = JSON.parse(localStorage.getItem('users')) || [];
     const currentDate = new Date().toLocaleDateString();
-
-    newUserData = {
-        username: usernameInput.value,
-        password: passwordInput.value,
-        accountDate: currentDate,
-        gamesPlayed: 0,
-        moneyWon: 0.0,
-        games: [0]
+    alreadyExists = false;
+    if (users) {
+        for (var i = 0; i < users.length; i++) {
+            console.log(users[i].username)
+            console.log(usernameInput)
+        if (users[i].username == usernameInput.value) {
+            alert("Username already taken. Please try another one.");
+            alreadyExists = true;
+        }
+        }
     }
+    if (!alreadyExists) {
+        
+        newUserData = {
+            username: usernameInput.value,
+            password: passwordInput.value,
+            accountDate: currentDate,
+            gamesPlayed: 0,
+            moneyWon: 0.0,
+            games: [0]
+        }
 
-    users.push(newUserData);
-    localStorage.setItem('users', JSON.stringify(users));
-    localStorage.setItem('currUser', usernameInput.value);
-    
-    window.location.href = "profile.html";
+        users.push(newUserData);
+        localStorage.setItem('users', JSON.stringify(users));
+        localStorage.setItem('currUser', usernameInput.value);
+        
+        window.location.href = "profile.html";
+    }
     
   } else {
     alert('Please fill out both username and password fields.');
